requirements.txt
python-dotenv
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
scikit-learn
numpy
python-multipart